Work around a standards compliance issue with pidgin.
Pidgin sends the initial NICK and USER message out of order. I'm
probably going to be stuck with more Postel's law than I'd like. We
should file a bug in pidgin.

I've been using bitlbee for testing, just because it's handy. Initial
connection seems to work with pidgin now, but reconnects fail. Bitlbee
complains about recieving a NICK message after the initial login. My
understanding of the spec was that you could use a NICK message to
change your nick, so maybe this is a bitlbee issue, but I should
investigate.