A few fixes, cleanup, and more logging
Wrapped the dropClient/reset calls in writemessage failures into their
own method. This reduces repetition, but it also makes it easy for us to
print debug info for every message that goes over the wire.

This also fixes a couple cases where we were erroneously forwarding USER
messages from the client's reconnecting handshake to the server, which
caused problems since it thinks we're already logged in.

Pidgin still isn't reconnecting successfully. My suspicion is it's
waiting for the rest of the welcome/info/maybe MOTD messages from the
server; shouldn't be hard to send those.