Use sql storage backend on non-sandstorm.
This gives us the ability to not lose already logged messages if the
server gets shut down. We haven't enabled it in sandstorm yet because
we need to do a little bit of work to link sqlite into the executable;
right now we rely on being able to cross compile and omit cgo stuff.
That's next on the agenda.

Also note that we're being a little sloppy re: maintaining the list of
users present in channels, and this will probably make that a little bit
worse.