Tweak PING timeout handling
With the previous implementation, if handling another event takes long
enough, we could end up blocking for long enough that we miss the PING
deadline and go straight to the DROP deadline. We now track whether
we've actually sent a PING.

The situation could conceivably come up if we're replaying a large log,
for example.